pos Application

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -1,8 +1,11 @@
 package com.ust.pos.user;
 
 import com.ust.pos.dto.UserDto;
+import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -11,35 +14,86 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/user")
 public class UserController {
 
+    public static final String MESSAGE = "message";
+    public static final String ROLES = "roles";
+    public static final String USER_USER = "user/user";
     @Autowired
     private UserService userService;
 
+    @Autowired
+    public RoleService roleService;
+
     @GetMapping("/list")
-    public String home(Model model) {
+    public String list(Model model) {
         model.addAttribute("users", userService.findAll());
         return "user/list";
     }
 
     @GetMapping("/get")
     public String update(Model model, @RequestParam String username) {
+
         UserDto response = userService.findByUserName(username);
+
         model.addAttribute("user", response);
-        return "user/user";
+        model.addAttribute(ROLES, roleService.findAll());
+
+        return USER_USER;
     }
 
     @PostMapping("/update")
-    public String updatePost(Model model, @ModelAttribute UserDto userDto) {
-        UserDto response = userService.update(userDto);
+    public String updatePost(Model model,
+                             @ModelAttribute UserDto userDto,
+                             @RequestParam String oldUsername) {
+
+        UserDto existingUser = userService.findByUserName(oldUsername);
+
+        if (existingUser == null) {
+            model.addAttribute(MESSAGE, "User not found.");
+            model.addAttribute(ROLES, roleService.findAll());
+            return USER_USER;
+        }
+
+        if (!oldUsername.equalsIgnoreCase(userDto.getUsername())) {
+
+            UserDto emailCheck =
+                    userService.findByUserName(userDto.getUsername());
+
+            if (emailCheck != null) {
+                model.addAttribute(MESSAGE,
+                        "❌ Email already exists. Please use a new email.");
+                model.addAttribute("user", userDto);
+                model.addAttribute(ROLES, roleService.findAll());
+                return USER_USER;
+            }
+        }
+
+        UserDto response =
+                userService.update(oldUsername, userDto);
+
         if (!response.isSuccess()) {
-            model.addAttribute("message", response.getMessage());
-            return "user/user";
+            model.addAttribute(MESSAGE, response.getMessage());
+            model.addAttribute("user", userDto);
+            model.addAttribute(ROLES, roleService.findAll());
+            return USER_USER;
         }
         return "redirect:/user/list";
     }
-
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
-        return "user/user";
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUser = authentication.getName();
+            if (loggedInUser != null) {
+
+                userService.delete(username);
+
+                if (loggedInUser.equals(username)) {
+                    SecurityContextHolder.clearContext();
+                    return "redirect:/login";
+                }
+            }
+        }
+        return "redirect:/user/list";
     }
 }
